+++ facenet/callbacks.py
@@ -1,0 +1,28 @@
+# coding: utf-8
+__author__ = 'Ruslan N. Kosarev'
+
+from loguru import logger
+
+import tensorflow as tf
+
+from facenet import facenet_tf2 as facenet
+from facenet import statistics
+
+
+class ValidateCallback(tf.keras.callbacks.Callback):
+    def __init__(self, model, dataset, every_n_epochs, max_nrof_epochs, config):
+        super().__init__()
+        self.model = model
+        self.dataset = dataset
+        self.config = config
+        self.every_n_epochs = every_n_epochs
+        self.max_nrof_epochs = max_nrof_epochs
+
+    def on_epoch_end(self, epoch, logs=None):
+        epoch1 = epoch + 1
+
+        if epoch1 % self.every_n_epochs == 0 or epoch1 == self.max_nrof_epochs:
+            logger.info(f'perform validation for epoch {epoch1}')
+
+            embeddings, labels = facenet.evaluate_embeddings(self.model, self.dataset)
+            statistics.FaceToFaceValidation(embeddings, labels, self.config.validate)
